pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "py-tgcalls>=2.1.1",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.0",
    "tgcrypto>=1.2.5",
    "yt-dlp>=2025.3.31",
]
